chore: enhance Flags SDK with caching configuration and add example application
